successful install

+++ script/lerna-toggle-package-manager.js
@@ -39,15 +39,19 @@ if(engine === 'npm') {
     hoist: true,
     ignoreScripts: true
   };
-  fs.unlink(resolveApp('yarn.lock'))
+  if(fs.existsSync(resolveApp('yarn.lock'))){
+    fs.unlink(resolveApp('yarn.lock'))
+  }
 } else if(engine === 'yarn') {
   packageJSON.private = true;
-  packageJSON.workspaces = "packages/*"
+  packageJSON.workspaces = ["packages/*"];
 
   lernaJSON.npmClient = "yarn";
   lernaJSON.useWorkspaces = true;
   delete lernaJSON.command.bootstrap;
-  fs.unlink(resolveApp('package-lock.json'))
+  if(fs.existsSync(resolveApp('package-lock.json'))){
+    fs.unlink(resolveApp('package-lock.json'))
+  }
 } else {
   program.outputHelp()
   return;
